+++ resources/js/components/Layout/Default.vue
@@ -9,7 +9,3 @@
         />
     </div>
 </template>
-
-<script setup>
-import Toast from 'primevue/toast';
-</script>