+++ resources/js/components/Nav/ProfileMenu.vue
@@ -24,6 +24,7 @@
                 <template #start>
                     <button
                         class="relative overflow-hidden w-full p-link flex items-center p-2 pl-3 text-surface-800 dark:text-white hover:bg-surface-200 dark:hover:bg-surface-600 border-noround"
+                        @click="userEditModal = true"
                     >
                         <div>
                             <Avatar
@@ -68,11 +69,29 @@
             </Link>
         </div>
     </div>
+    <Dialog v-model:visible="userEditModal" header="Edit Profile" :style="{ width: '25rem' }">
+        <span class="text-surface-500 dark:text-surface-400 block mb-8">Update your information.</span>
+        <div class="flex items-center gap-4 mb-4">
+            <label for="username" class="font-semibold w-24">Username</label>
+            <InputText id="username" class="flex-auto" autocomplete="off" />
+        </div>
+        <div class="flex items-center gap-4 mb-8">
+            <label for="email" class="font-semibold w-24">Email</label>
+            <InputText id="email" class="flex-auto" autocomplete="off" />
+        </div>
+        <div class="flex justify-end gap-2">
+            <Button type="button" label="Cancel" severity="secondary" @click="visible = false"></Button>
+            <Button type="button" label="Save" @click="visible = false"></Button>
+        </div>
+    </Dialog>
 </template>
 
 <script setup>
-const page = usePage()
-const user = computed(() => page.props?.user)
+const { modal } = useModal();
+
+const userEditModal = modal('UserEditModal')
+
+const user = usePageProp('user');
 const menu = ref();
 const items = ref([
     {