+++ resources/js/pages/Admin/CreateUser.vue
@@ -78,7 +78,7 @@
 </template>
 
 <script setup>
-import { useToast } from 'primevue/usetoast';
+// import { useToast } from 'primevue/usetoast';
 const form = useForm({
     name: null,
     email: null,