+++ resources/js/pages/Admin/Users.vue
@@ -176,7 +176,7 @@
 </template>
 
 <script setup>
-import { useToast } from 'primevue/usetoast';
+// import { useToast } from 'primevue/usetoast';
 const toast = useToast();
 const props = defineProps({
     users: {