+++ CHANGELOG.md
@@ -1,3 +1,14 @@
 # Changelog
 
 Any project using this template should follow the changelog here if its relevant to make the needed updates to be compatible. 
+
+# v2.0.0 - 04/26/2025
+
+This update centralizes the base components into atlas ui repo
+
+- Install package `npm install github:tmarois/atlas-ui#semver:^1.0.0`
+- Updated `tailwind.config.js`
+- Updated `vite.config.js` (new components and composables for autoloading)
+- Updated `resources/js/setup.js`
+- Updated `resources/css/app.css`
+- Deleted all `/components/_volt/~` (now loading from custom atlas-ui package)